+++ src/main/java/com/drmangotea/tfmg/content/machinery/oil_processing/OilHammerItem.java
@@ -1,12 +1,10 @@
 package com.drmangotea.tfmg.content.machinery.oil_processing;
 
 import com.drmangotea.tfmg.base.lang.TFMGLang;
-import com.drmangotea.tfmg.content.world.LevelDataHandler;
-import com.drmangotea.tfmg.content.world.resevoir.FluidReservoirs;
-import com.drmangotea.tfmg.registry.TFMGBlocks;
+import com.drmangotea.tfmg.content.world.resevoir.FluidReservoir;
+import com.drmangotea.tfmg.registry.TFMGDataAttachments;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
@@ -24,28 +22,15 @@ public class OilHammerItem extends Item {
         BlockPos pos = context.getClickedPos();
         Player player = context.getPlayer();
 
-        FluidReservoirs reservoirs = null;
-
-        if (level instanceof ServerLevel serverLevel) {
-            reservoirs = LevelDataHandler.getFluidReservoirs(serverLevel);
+        if (!level.getChunk(pos).hasData(TFMGDataAttachments.FLUID_RESERVOIR)) {
+            return InteractionResult.sidedSuccess(level.isClientSide());
         }
+        FluidReservoir reservoir = level.getChunk(pos).getData(TFMGDataAttachments.FLUID_RESERVOIR);
 
-        if (reservoirs != null) {
-            for(int i = 0; i < 300; i++) {
-                BlockPos posToCheck = pos.below(i);
-                if(level.getBlockState(posToCheck).is(TFMGBlocks.OIL_DEPOSIT.get())) {
-                    if(reservoirs.getReservoirFor(posToCheck) == null)
-                        return InteractionResult.SUCCESS;
-                    int oilReserves = reservoirs.getReservoirFor(posToCheck).getReserves();
-
-                    if (level.isClientSide && player != null)
-                        player.displayClientMessage(TFMGLang.translateDirect("oil_hammer.reserves", oilReserves)
-                                .withStyle(ChatFormatting.YELLOW), true);
-
-                    return InteractionResult.SUCCESS;
-                }
-            }
-        }
+        int oilReserves = reservoir.getReserves();
+        if (level.isClientSide && player != null)
+            player.displayClientMessage(TFMGLang.translateDirect("oil_hammer.reserves", oilReserves)
+                    .withStyle(ChatFormatting.YELLOW), true);
 
         return InteractionResult.SUCCESS;
     }